+++ packages/react/src/receipt-blocks/itemized-car-rental/index.ts
@@ -1,1 +1,2 @@
+export * from "./itemized-car-rental-2.1.0";
 export * from "./itemized-car-rental";