+++ packages/schema/lts.ts
@@ -1,3 +1,3 @@
 /** This is the source of truth for currently supported Schema Versions by this library */
 
-module.exports = ["1.11.0", "2.0.0", "2.1.0"];
+module.exports = ["1.11.0", "2.0.0", "2.1.0", "2.1.1"];