+++ packages/schema/src/index.ts
@@ -2,7 +2,7 @@
 export * as lts from "./lts";
 
 /* Always re-export the latest version directly */
-export * from "./lts/2.1.0";
+export * from "./lts/2.1.1";
 
 /* And finally export a list of LTS versions */
-export const LTS_VERSIONS = ["1.11.0", "2.0.0", "2.1.0"];
+export const LTS_VERSIONS = ["1.11.0", "2.0.0", "2.1.0", "2.1.1"];