+++ packages/schema/src/lts/index.ts
@@ -2,3 +2,4 @@
 export * as v1_11_0 from "./1.11.0";
 export * as v2_0_0 from "./2.0.0";
 export * as v2_1_0 from "./2.1.0";
+export * as v2_1_1 from "./2.1.1";