+++ packages/schema/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @versa/schema
 
+## 2.1.1
+
+### Patch Changes
+
+- patch update to 2.1.1
+
 ## 2.1.0
 
 ### Minor Changes